channelcoord: move kv restore shell test
Move RestoreChannelShells coverage from channeldb into channelcoord because the operation coordinates channel-state restoration with link-node creation.

The moved test builds the KV coordinator fixture directly and checks the persisted link-node address by its normalized string form.
