+++ channelcoord/kv_coordinator_test.go
@@ -2,6 +2,8 @@ package channelcoord_test
 
 import (
 	"bytes"
+	"math"
+	"math/rand"
 	"net"
 	"sync/atomic"
 	"testing"
@@ -167,6 +169,72 @@ func createKVCoordinatorTestChannel(t *testing.T,
 	}
 }
 
+func genRandomKVChannelShell() (*chanstate.ChannelShell, error) {
+	var testPriv [32]byte
+	if _, err := rand.Read(testPriv[:]); err != nil {
+		return nil, err
+	}
+
+	_, pub := btcec.PrivKeyFromBytes(testPriv[:])
+
+	var chanPoint wire.OutPoint
+	if _, err := rand.Read(chanPoint.Hash[:]); err != nil {
+		return nil, err
+	}
+
+	chanPoint.Index = uint32(rand.Intn(math.MaxUint16))
+
+	var shaChainPriv [32]byte
+	if _, err := rand.Read(shaChainPriv[:]); err != nil {
+		return nil, err
+	}
+
+	revRoot, err := chainhash.NewHash(shaChainPriv[:])
+	if err != nil {
+		return nil, err
+	}
+	shaChainProducer := shachain.NewRevocationProducer(*revRoot)
+
+	commitParams := chanstate.CommitmentParams{
+		CsvDelay: uint16(rand.Int63()),
+	}
+
+	channel := &chanstate.OpenChannel{
+		ChainHash:       *revRoot,
+		FundingOutpoint: chanPoint,
+		ShortChannelID: lnwire.NewShortChanIDFromInt(
+			uint64(rand.Int63()),
+		),
+		IdentityPub: pub,
+		LocalChanCfg: chanstate.ChannelConfig{
+			CommitmentParams: commitParams,
+			PaymentBasePoint: keychain.KeyDescriptor{
+				KeyLocator: keychain.KeyLocator{
+					Family: keychain.KeyFamily(
+						rand.Int63(),
+					),
+					Index: uint32(rand.Int63()),
+				},
+			},
+		},
+		RemoteCurrentRevocation: pub,
+		IsPending:               false,
+		RevocationStore:         shachain.NewRevocationStore(),
+		RevocationProducer:      shaChainProducer,
+	}
+	channel.SetChannelStatusForStore(
+		chanstate.ChanStatusDefault | chanstate.ChanStatusRestored,
+	)
+
+	return &chanstate.ChannelShell{
+		NodeAddrs: []net.Addr{&net.TCPAddr{
+			IP:   net.ParseIP("127.0.0.1"),
+			Port: 18555,
+		}},
+		Chan: channel,
+	}, nil
+}
+
 // TestRepairLinkNodes tests that RepairLinkNodes identifies and repairs
 // missing link nodes for channels that exist in the database.
 func TestRepairLinkNodes(t *testing.T) {
@@ -222,4 +290,55 @@ func TestRepairLinkNodes(t *testing.T) {
 	)
 	require.NoError(t, err, "repaired link node should exist")
 	require.Equal(t, wire.TestNet3, repairedLinkNode.Network)
+}
+
+// TestRestoreChannelShells tests that we're able to insert a partially
+// populated channel to disk and create its related link-node metadata.
+func TestRestoreChannelShells(t *testing.T) {
+	t.Parallel()
+
+	fixture := makeKVCoordinatorTestFixture(t)
+
+	channelShell, err := genRandomKVChannelShell()
+	require.NoError(t, err, "unable to gen channel shell")
+
+	err = fixture.coordinator.RestoreChannelShells(channelShell)
+	require.NoError(t, err, "unable to restore channel shell")
+
+	nodeChans, err := fixture.chanStore.FetchOpenChannels(
+		channelShell.Chan.IdentityPub,
+	)
+	require.NoError(t, err, "unable find channel")
+	require.Len(t, nodeChans, 1)
+
+	channel := nodeChans[0]
+	_, err = channel.UpdateCommitment(nil, nil)
+	require.ErrorIs(t, err, chanstate.ErrNoRestoredChannelMutation)
+
+	err = channel.AppendRemoteCommitChain(nil)
+	require.ErrorIs(t, err, chanstate.ErrNoRestoredChannelMutation)
+
+	err = channel.AdvanceCommitChainTail(nil, nil, 0, 1)
+	require.ErrorIs(t, err, chanstate.ErrNoRestoredChannelMutation)
+
+	require.Equal(
+		t, channelShell.Chan.FundingOutpoint,
+		channel.FundingOutpoint,
+	)
+	require.True(t, channel.HasChanStatus(chanstate.ChanStatusRestored))
+
+	_, err = fixture.chanStore.FetchChannel(
+		channelShell.Chan.FundingOutpoint,
+	)
+	require.NoError(t, err, "unable to fetch channel")
+
+	linkNode, err := fixture.linkNodeStore.FetchLinkNode(
+		channelShell.Chan.IdentityPub,
+	)
+	require.NoError(t, err, "unable to fetch link node")
+	require.Len(t, linkNode.Addresses, len(channelShell.NodeAddrs))
+	require.Equal(
+		t, channelShell.NodeAddrs[0].String(),
+		linkNode.Addresses[0].String(),
+	)
 }
